test: move reth-primitives tx tests
